Combine conditions interfaces
